% Generated by roxygen2: do not edit by hand
% Please edit documentation in LinearRegression/linearRegression.R
\name{.linregMakePrettyNames}
\alias{.linregMakePrettyNames}
\title{Change R names into pretty names, e.g., PredictorLevel ->}
\usage{
.linregMakePrettyNames(info)
}
\arguments{
\item{info}{either an lm object, a list where each sub element $fit contains an lm object, or the result of .linregGetParametersAndLevels.}
}
\description{
Change R names into pretty names, e.g., PredictorLevel ->
}
\details{
For example "PredictorLevel" becomes "Predictor" Level
}
