% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{ImputationInternal}
\alias{ImputationInternal}
\title{Multiply impute missing data with MICE}
\usage{
ImputationInternal(jaspResults, dataset, options)
}
\description{
Multiply impute missing data with MICE
}
