% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingDataImputation.R
\name{MissingDataImputation}
\alias{MissingDataImputation}
\title{Multiply impute missing data with MICE}
\usage{
MissingDataImputation(jaspResults, dataset, options)
}
\description{
Multiply impute missing data with MICE
}
